requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.105.0
h11==0.14.0
idna==3.6
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
pydantic==2.5.2
pydantic_core==2.14.5
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
